# swagger.yml
openapi: 3.0.2
...
info:
  title: Ercole
  version: 2.0.0
  license:
    name: GNU General Public License v3.0 or later
    url: "https://www.gnu.org/licenses/gpl-3.0-standalone.html"
  description: API used to talk to ercole services
  contact:
    name: Issues support
    url: "https://github.com/ercole-io/ercole"
servers:
  - description: Api service on dev.ercole.io
    url: "https://dev.ercole.io/api"
  - description: Chart service on dev.ercole.io
    url: "https://dev.ercole.io/chart"
  - description: Data service on dev.ercole.io
    url: "https://dev.ercole.io/data"
  - description: Alert service on dev.ercole.io
    url: "https://dev.ercole.io/alert"
  - description: Repo service on dev.ercole.io
    url: "https://dev.ercole.io/repo"
  - description: Data service
    url: "http://localhost:11111"
  - url: "http://localhost:11112"
    description: Alert service
  - url: "http://localhost:11113"
    description: Api service
  - url: "http://localhost:11116"
    description: Chart service
...
components:
  securitySchemes:
    basicAuthDataService:
      type: http
      scheme: basic
      description: access to data-service APIs
    basicAuthAlertService:
      type: http
      scheme: basic
      description: access to alert-service APIs
    basicAuthApiService:
      type: http
      scheme: basic
      description: access to api-service APIs
    tokenAuthApiService:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: access using the access token
  responses:
    error:
      description: Contains some informations about an error
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            properties:
              error:
                type: string
              message:
                type: string
              sourceFilename:
                type: string
              lineNumber:
                type: integer
            required:
              - error
              - message
              - sourceFilename
              - lineNumber
          examples:
            Test1:
              value:
                error: Invalid  status
                errorDescription: invalid status
                sourceFilename: /home/travis/go/src/github.com/ercole-io/ercole/api-service/controller/alerts_api.go
                lineNumber: 65
  schemas:
...
    HostDataInfo:
      type: object
      required:
        - hostname
        - environment
        - location
        - cpuModel
        - cpuCores
        - cpuThreads
        - socket
        - type
        - virtual
        - kernel
        - os
        - memoryTotal
        - swapTotal
        - oracleCluster
        - veritasCluster
        - sunCluster
        - aixCluster
      properties:
        hostname:
          type: string
          pattern: '^(([a-zA-Z0-9]|[a-zA-Z0-9][a-zA-Z0-9\-]*[a-zA-Z0-9])\.)*([A-Za-z0-9]|[A-Za-z0-9][A-Za-z0-9\-]*[A-Za-z0-9])$'
        environment:
          type: string
        location:
          type: string
        cpuModel:
          type: string
        cpuCores:
          type: integer
        cpuThreads:
          type: integer
        socket:
          type: integer
        type:
          type: string
        virtual:
          type: boolean
        kernel:
          type: string
        os:
          type: string
        memoryTotal:
          type: number
        swapTotal:
          type: number
        oracleCluster:
          type: boolean
        veritasCluster:
          type: boolean
        sunCluster:
          type: boolean
        aixCluster:
          type: boolean
        clusterwareVersion:
          type: string
    Filesystems:
      type: array
      items:
        type: object
        properties:
          filesystem:
            type: string
          fsType:
            type: string
          size:
            type: string
          used:
            type: string
          usedPerc:
            type: string
          mountedOn:
            type: string
        required:
          - filesystem
          - fsType
          - size
          - used
          - usedPerc
          - mountedOn
...
    Cluster:
      type: object
      properties:
        Name:
          type: string
        Environment:
          type: string
          nullable: true
        Location:
          type: string
          nullable: true
        Hostname:
          type: string
          nullable: true
        HostnameAgentVirtualization:
          type: string
          nullable: true
        VirtualizationNodes:
          nullable: true
          anyOf:
            - type: string
              nullable: true
            - type: array
              items:
                type: string
              uniqueItems: true
        PhysicalServerModelNames:
          nullable: true
          anyOf:
            - type: string
              nullable: true
            - type: array
              items:
                type: string
              uniqueItems: true
        VirtualizationNodesCount:
          nullable: true
          type: integer
        VirtualizationNodesStats:
          nullable: true
          type: object
          required:
            - VirtualizationNode
            - TotalVMsCount
            - TotalVMsWithErcoleAgentCount
            - TotalVMsWithoutErcoleAgentCount
          properties:
            VirtualizationNode:
              type: string
            TotalvMsCount:
              type: integer
            TotalvMsWithErcoleAgentCount:
              type: integer
            TotalvMsWithoutErcoleAgentCount:
              type: integer
        Type:
          type: string
        cpu:
          type: integer
        Sockets:
          type: integer
        vmsCount:
          type: integer
          nullable: true
        vmsErcoleAgentCount:
          type: integer
          nullable: true
        vms:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              clusterName:
                type: string
              hostname:
                type: string
              cappedCPU:
                type: boolean
              virtualizationNode:
                type: string
              physicalServerModelName:
                type: string
              isErcoleInstalled:
                type: boolean
            required:
              - name
              - clusterName
              - hostname
              - cappedCPU
              - virtualizationNode
      required:
        - name
        - type
        - cpu
        - sockets
        - vms
...
    HostData:
      description: A hostdata from FE
      type: object
      properties:
        hostname:
          type: string
        environment:
          type: string
        location:
          type: string
        version:
          type: string
        hostDataSchemaVersion:
          type: integer
        databases:
          type: string
        schemas:
          type: string
          nullable: true
        info:
          $ref: "#/components/schemas/HostDataInfo"
        os:
          type: string
          nullable: true
        kernel:
          type: string
          nullable: true
        oracleCluster:
          type: boolean
          nullable: true
        sunCluster:
          type: boolean
          nullable: true
        veritasCluster:
          type: boolean
          nullable: true
        virtual:
          type: boolean
          nullable: true
        type:
          type: string
          nullable: true
        cpuThreads:
          type: integer
          nullable: true
        cpuCores:
          type: integer
          nullable: true
        socket:
          type: integer
          nullable: true
        memTotal:
          type: integer
          nullable: true
        swapTotal:
          type: integer
          nullable: true
        cpuModel:
          type: string
          nullable: true
        physicalServerName:
          type: string
          nullable: true
        virtualServerName:
          type: string
          nullable: true
        virtualizationTechnology:
          type: string
          nullable: true
        dbInstanceName:
          type: string
          nullable: true
        pluggableDatabaseName:
          type: string
          nullable: true
        connectString:
          type: string
          nullable: true
        productVersion:
          type: string
          nullable: true
        productEdition:
          type: string
          nullable: true
        features:
          type: object
          properties:
            oracle:
              type: object
              properties:
                databases:
                  type: array
                  items:
                    type: object
                    properties:
                      "":
                        type: string
                unlistedRunningDatabases:
                  type: array
                  items:
                    type: string
        racNodeNames:
          type: string
          nullable: true
        processorModel:
          type: string
          nullable: true
        processors:
          type: integer
          nullable: true
        coresPerProcessor:
          type: integer
          nullable: true
        threadsPerCore:
          type: integer
          nullable: true
        processorSpeed:
          type: string
          nullable: true
        serverPurchaseDate:
          type: string
          nullable: true
        operatingSystem:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        physicalCores:
          type: integer
          nullable: true
        history:
          type: array
          items:
            $ref: "#/components/schemas/HistoryEntry"
        extra:
          type: object
          required:
            - filesystems
          properties:
            filesystems:
              $ref: "#/components/schemas/Filesystems"
            databases:
              nullable: true
              type: array
              items:
                $ref: "#/components/schemas/Database"
            clusters:
              nullable: true
              type: array
              items:
                $ref: "#/components/schemas/Cluster"
        cluster:
          type: string
          nullable: true
        virtualizationNode:
          type: string
          nullable: true
        consumptions:
          type: array
          items:
            type: object
            properties:
              timeStart:
                type: string
                format: date-time
              timeEnd:
                type: string
                format: date-time
              cpuAvg:
                type: number
        isMissingDB:
          type: array
          items:
            type: string
        memorySum:
          type: number
        memorySumPercentage:
          type: number
        memorySumFlag:
          type: string
      required:
        - hostname
        - environment
        - location
        - version
        - hostDataSchemaVersion
        - info
        - extra
...
    PageMetadata:
      type: object
      required:
        - empty
        - first
        - last
        - number
        - size
        - totalElements
        - totalPages
      properties:
        empty:
          type: boolean
        first:
          type: boolean
        last:
          type: boolean
        number:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
    HistoryEntry:
      type: object
      required:
        - _id
        - createdAt
      properties:
        _id:
          type: string
        createdAt:
          type: string
        totalDailycpuUsage:
          type: number
...
  parameters:
    search:
      in: query
      name: search
      required: false
      description: filter the result using search terms
      allowEmptyValue: true
      example: foobar
      schema:
        type: string
        readOnly: true
        example: foobar
    sort-by:
      in: query
      name: sort-by
      required: false
      description: sort the result by a field
      allowEmptyValue: true
      example: foobar
      schema:
        type: string
        readOnly: true
        example: CreatedAt
    sort-desc:
      schema:
        type: boolean
      in: query
      name: sort-desc
      description: |
        Sort the field descendently instead of ascendently
      allowEmptyValue: true
    page:
      schema:
        type: integer
      in: query
      name: page
      description: Number of the page
      allowEmptyValue: true
    size:
      schema:
        type: integer
      in: query
      name: size
      description: Size of the page
      allowEmptyValue: true
    location:
      schema:
        type: string
        example: Italy
      in: query
      name: location
      description: Filter by location
      allowEmptyValue: true
    environment:
      schema:
        type: string
        example: TST
      in: query
      name: environment
      description: Filter by environment
      allowEmptyValue: true
    older-than:
      schema:
        type: string
        format: date-time
        example: "2020-05-05T08:23:40.000Z"
      in: query
      name: older-than
      description: Filter until the date
      allowEmptyValue: true
...
paths:
...
  /hosts:
    get:
      tags:
        - api-service
        - fe-user
      operationId: SearchHosts
      summary: Search a list of hosts
      description: Get a list of hosts filtered using various search terms and various params. Can also generate a XLSX file
      parameters:
        - in: query
          name: mode
          description: change the scheme of the output
          allowEmptyValue: true
          example: full
          schema:
            type: string
            readOnly: true
            enum:
              - full
              - summary
              - lms
              - mhd
        - $ref: "#/components/parameters/search"
        - in: query
          name: hostname
          description: filter by hostname
          allowEmptyValue: true
          schema:
            type: string
        - in: query
          name: database
          description: filter by a database name
          allowEmptyValue: true
          schema:
            type: string
        - in: query
          name: technology
          description: filter by a technology
          allowEmptyValue: true
          schema:
            type: string
        - in: query
          name: hardware-abstraction-technology
          description: filter by the HA technology
          allowEmptyValue: true
          schema:
            type: string
        - in: query
          name: cluster
          description: filter by the name of virtualization cluster. use 'NULL' for searching host that aren't part of a virtualization cluster
          allowEmptyValue: true
          schema:
            type: string
        - in: query
          name: physical-host
          description: filter by the name of physical host on which the host runs
          allowEmptyValue: true
          schema:
            type: string
        - in: query
          name: operating-system
          description: filter by the operating system
          allowEmptyValue: true
          schema:
            type: string
        - in: query
          name: kernel
          description: filter by the kernel
          allowEmptyValue: true
          schema:
            type: string
        - in: query
          name: memory-total-lte
          description: filter the hosts with the Info.MemoryTotal less than memory-total-lte value
          allowEmptyValue: true
          schema:
            type: number
        - in: query
          name: memory-total-gte
          description: filter the hosts with the Info.MemoryTotal less than memory-total-gte value
          allowEmptyValue: true
          schema:
            type: number
        - in: query
          name: swap-total-lte
          description: filter the hosts with the Info.SwapTotal less than swap-total-lte value
          allowEmptyValue: true
          schema:
            type: number
        - in: query
          name: swap-total-gte
          description: filter the hosts with the Info.SwapTotal less than swap-total-gte value
          allowEmptyValue: true
          schema:
            type: number
        - in: query
          name: is-member-of-cluster
          description: filter the host by operating systems cluster membership
          allowEmptyValue: true
          schema:
            type: boolean
        - in: query
          name: cpu-model
          description: filter by CPU model
          allowEmptyValue: true
          schema:
            type: boolean
        - in: query
          name: cpu-cores-lte
          description: filter the hosts with the Info.CPUCores less than cpu-cores-lte value
          allowEmptyValue: true
          schema:
            type: number
        - in: query
          name: cpu-cores-gte
          description: filter the hosts with the Info.CPUCores less than cpu-cores-gte value
          allowEmptyValue: true
          schema:
            type: number
        - in: query
          name: cpu-threads-lte
          description: filter the hosts with the Info.CPUThreads less than cpu-threads-lte value
          allowEmptyValue: true
          schema:
            type: number
        - in: query
          name: cpu-threads-gte
          description: filter the hosts with the Info.CPUThreads less than cpu-threads-gte value
          allowEmptyValue: true
          schema:
            type: number
        - $ref: "#/components/parameters/sort-by"
        - $ref: "#/components/parameters/sort-desc"
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/size"
        - $ref: "#/components/parameters/location"
        - $ref: "#/components/parameters/environment"
        - $ref: "#/components/parameters/older-than"
      responses:
        "200":
          description: Result of the search
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    description: full paged
                    properties:
                      content:
                        type: array
                        items:
                          $ref: "#/components/schemas/HostData"
                      metadata:
                        $ref: "#/components/schemas/PageMetadata"
                    required:
                      - content
                      - metadata
                  - type: array
                    description: full non paged
                    items:
                      $ref: "#/components/schemas/HostData"
                  - type: object
                    description: summary paged
                    properties:
                      content:
                        type: array
                        items:
                          $ref: "#/components/schemas/HostData"
                      metadata:
                        $ref: "#/components/schemas/PageMetadata"
                    required:
                      - content
                      - metadata
                  - type: array
                    description: summary non paged
                    items:
                      $ref: "#/components/schemas/HostData"
                  - type: object
                    description: lms paged
                    properties:
                      content:
                        type: array
                        items:
                          $ref: "#/components/schemas/HostData"
                      metadata:
                        $ref: "#/components/schemas/PageMetadata"
                    required:
                      - content
                      - metadata
                  - type: array
                    description: lms non paged
                    items:
                      $ref: "#/components/schemas/HostData"
            ? application/vnd.oracle.lms+vnd.openxmlformats-officedocument.spreadsheetml.sheet
            : {}
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              {}
        "401":
          $ref: "#/components/responses/error"
        "422":
          $ref: "#/components/responses/error"
        "500":
          $ref: "#/components/responses/error"
  "/hosts/{hostname}":
    parameters:
      - in: path
        name: hostname
        schema:
          type: string
        required: true
        description: hostname of the requested host
    get:
      tags:
        - api-service
        - fe-user
        - read
      operationId: GetHost
      summary: Get a host
      description: Get a host
      parameters:
        - $ref: "#/components/parameters/older-than"
      responses:
        "200":
          description: Result of the search
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HostData"
    delete:
      tags:
        - api-service
        - fe-user
        - write
      operationId: DismissHost
      summary: Dismiss a host
      description: Dismiss a host
      responses:
        "200":
          description: The host is dismissed
          content: {}
        "401":
          $ref: "#/components/responses/error"
        "404":
          description: The host is not found
        "422":
          $ref: "#/components/responses/error"
        "500":
          $ref: "#/components/responses/error"
...
  /hosts/locations:
    get:
      tags:
        - api-service
        - fe-user
        - read
      operationId: ListLocations
      summary: List users' locations
      description: List the users' locations
      parameters: []
      responses:
        "200":
          description: List of the locations
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                uniqueItems: true
        "401":
          $ref: "#/components/responses/error"
        "422":
          $ref: "#/components/responses/error"
        "500":
          $ref: "#/components/responses/error"
  /hosts/environments:
    get:
      tags:
        - api-service
        - fe-user
        - read
      operationId: ListEnvironments
      summary: List the environments
      description: List the environments
      parameters:
        - $ref: "#/components/parameters/location"
        - $ref: "#/components/parameters/environment"
        - $ref: "#/components/parameters/older-than"
      responses:
        "200":
          description: List of the environments
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                uniqueItems: true
        "401":
          $ref: "#/components/responses/error"
        "422":
          $ref: "#/components/responses/error"
        "500":
          $ref: "#/components/responses/error"
...
security:
  - basicAuthApiService: []
  - tokenAuthApiService: []